+++ openeo_driver/server.py
@@ -1,4 +1,3 @@
-import datetime
 import logging
 import logging.config
 from typing import List